Changes type of WeekdayCount from int to short

+++ src/bedtimes.h
@@ -1,5 +1,5 @@
 #pragma once
 
-const int WeekdayCount = 7;
+const short WeekdayCount = 7;
 
 bool ParseBedtimes(const char *Bedtimes, char *Result[]);
